Make SafarisList responsive
