+++ src/components/Overlay.js
@@ -1,7 +1,20 @@
 import PropTypes from 'prop-types';
+import { useEffect } from 'react';
 
 function Overlay({ handleClick }) {
-  return <div className="fixed top-0 bottom-0 left-0 right-0 z-10 bg-[#a0afa033] backdrop-blur-sm duration-100" onClick={handleClick} role="presentation" />;
+  useEffect(() => {
+    document.body.style.overflow = 'hidden';
+    return () => {
+      document.body.style.overflow = 'scroll';
+    };
+  }, []);
+  return (
+    <div
+      className="fixed top-0 bottom-0 left-0 right-0 z-10 bg-[#707f7088] backdrop-blur-3xl duration-100"
+      onClick={handleClick}
+      role="presentation"
+    />
+  );
 }
 
 Overlay.propTypes = {
