+++ src/index.css
@@ -80,5 +80,5 @@
 
   /* *, ::before, ::after {
     outline: 1px solid red;
-  } */
-  
+  }
+   */
